fix: image generation wip
