+++ src/layouts/PostDetails.astro
@@ -17,7 +17,7 @@ const { title, author, description, ogImage, pubDatetime, readingTime, tags } = 
 
 const { Content } = await post.render();
 
-const ogUrl = new URL(ogImage ? ogImage : `${title}.png`, Astro.url.origin).href;
+const ogUrl = new URL(ogImage ? ogImage : `${post.slug}.png`, Astro.url.origin).href;
 ---
 
 <Layout title={title} author={author} description={description} ogImage={ogUrl}>
