--- src/pages/[ogTitle].svg.ts
@@ -1,19 +1,0 @@
-import { getCollection } from "astro:content";
-import generateOgImage from "@utils/generateOgImage";
-import type { APIRoute } from "astro";
-
-export const GET: APIRoute = async ({ params }) => new Response(await generateOgImage(params.ogTitle));
-
-const postImportResult = await getCollection(
-  "blog",
-  ({ data }) => import.meta.env.MODE === "development" || !data.draft
-);
-const posts = Object.values(postImportResult);
-
-export function getStaticPaths() {
-  return posts
-    .filter(({ data }) => !data.ogImage)
-    .map(({ data }) => ({
-      params: { ogTitle: data.title }
-    }));
-}
